vendorOS/requirements.txt
anthropic>=0.40.0
playwright>=1.44.0
beautifulsoup4>=4.12.0
requests>=2.31.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
pandas>=2.0.0
